requirements.txt
cairocffi==1.1.0
CairoSVG==2.4.2
cffi==1.13.2
cssselect2==0.2.2
defusedxml==0.6.0
googleads==3.8.0
httplib2==0.14.0
numpy==1.17.3
oauth2client==4.1.3
Pillow==6.2.1
praat-parselmouth==0.3.3
pyasn1==0.4.7
pyasn1-modules==0.2.7
PyAudio==0.2.11
pydicom==1.3.0
pydub==0.23.1
PySocks==1.7.1
python-magic==0.4.15
pytz==2015.7
PyYAML==5.1.2
rsa==4.0
scipy==1.3.1
six==1.13.0
SoundFile==0.10.2
stopit==1.1.1
suds-jurko==0.6
TextGrid==1.4
tinycss2==1.0.2
ttkthemes==2.3.0
urllib3==1.25.6
wav2vec==1.0.1
webencodings==0.5.1
xparser==0.0.4
